Don't verify database connection on `GetConnectionWithConfig`
Connection verification can take a very long time. Especially if the database is unavailable.
At the same time, the connection check in the `GetConnectionWithConfig` method is executed under the `b.connections` write lock.

In this case, we get a situation where create and revoke secrets makes it impossible to change the connections configuration:
we wait too many time in `connectionWriteHandler` for `b.connections` write lock and got `context.Timeout` error on `storeConfig`.